PWA-1384 Added helper function to iterate over all available IP addresses to find the local one.
