+++ lib/app/frontend/FrontendSettings.js
@@ -8,8 +8,8 @@
 const fsEx = require('fs-extra')
 const SETTINGS_FRONTEND_FILE = 'frontend.json'
 const path = require('path')
-const os = require('os')
 const logger = require('../../logger')
+const { getLocalIpAdresses } = require('../../utils/utils')
 
 class FrontendSettings {
   /**
@@ -99,7 +99,7 @@ class FrontendSettings {
    * Sets the startpage IP address.
    * @param {string} [startpageIp='0.0.0.0'] The network IP address
    */
-  async setStartpageIpAddress (startpageIp = os.networkInterfaces().en0[1].address) {
+  async setStartpageIpAddress (startpageIp = getLocalIpAdresses()) {
     const settings = (await this.loadSettings())
     settings.startpageIp = startpageIp
     await this._saveSettings(settings)
