+++ lib/app/frontend/FrontendSetup.js
@@ -6,11 +6,11 @@
  */
 
 const inquirer = require('inquirer')
-const os = require('os')
 const { green } = require('chalk')
 const logHelper = require('./LogHelper')
 const setupQuestions = require('./setupQuestions')
 const logger = require('../../logger')
+const { getLocalIpAdresses } = require('../../utils/utils')
 
 /**
  * The FrontendSetup class.
@@ -64,29 +64,6 @@ class FrontendSetup {
   }
 
   /**
-   * Returns the client's local IP address.
-   * @return {string}
-   */
-  getLocalIpAdresses () {
-    // Resolve the local area network ip.
-    const ifaces = os.networkInterfaces()
-    let localIp = '127.0.0.1'
-
-    Object.keys(ifaces).forEach((ifname) => {
-      ifaces[ifname].forEach((iface) => {
-        if (iface.family !== 'IPv4' || iface.internal !== false) {
-          // Skip over internal and non-ipv4 addresses
-          return
-        }
-
-        localIp = iface.address
-      })
-    })
-
-    return localIp
-  }
-
-  /**
    * Builds and returns the default configuration.
    * @return {Object}
    */
@@ -102,7 +79,7 @@ class FrontendSetup {
 
     return {
       ip: '0.0.0.0',
-      startpageIp: startpageIp || this.getLocalIpAdresses(),
+      startpageIp: startpageIp || getLocalIpAdresses(),
       port: port || 8080,
       apiPort: apiPort || 9666,
       hmrPort: hmrPort || 3000,
