+++ lib/utils/utils.js
@@ -2,6 +2,7 @@ const logger = require('../logger')
 const validator = require('@shopgate/extension-config-validator/validator.js')
 const fsEx = require('fs-extra')
 const path = require('path')
+const os = require('os')
 const { SETTINGS_FOLDER, EXTENSIONS_FOLDER, THEMES_FOLDER, TRUSTED_PIPELINES_FOLDER, PIPELINES_FOLDER } = require('../app/Constants')
 
 async function filterDir (dirPath, filterFunction) {
@@ -435,6 +436,29 @@ function getExtensionDirectoryName (extensionId = '') {
   return extensionId.replace(/\//g, '-')
 }
 
+/**
+ * Returns the client's local IP address.
+ * @return {string}
+ */
+function getLocalIpAdresses () {
+  // Resolve the local area network ip.
+  const ifaces = os.networkInterfaces()
+  let localIp = '127.0.0.1'
+
+  Object.keys(ifaces).forEach((ifname) => {
+    ifaces[ifname].forEach((iface) => {
+      if (iface.family !== 'IPv4' || iface.internal !== false) {
+        // Skip over internal and non-ipv4 addresses
+        return
+      }
+
+      localIp = iface.address
+    })
+  })
+
+  return localIp
+}
+
 module.exports = {
   resetProject,
   setProcessFile,
@@ -454,5 +478,6 @@ module.exports = {
   writeLocalPipelines,
   sleep,
   runSpinner,
-  getExtensionDirectoryName
+  getExtensionDirectoryName,
+  getLocalIpAdresses
 }
